[documentdb] Rename user command group to entra-user; clarify object-id identifier
Address Oliver Towers' PR review on the mongocluster user commands: the users
API only supports Microsoft Entra ID principals today, so the generic `user`
naming misleads customers about what is provisioned and what identifier to pass.

- Rename command group `mongocluster user` -> `mongocluster entra-user`.
- Rename the identifier arg `--user-name` -> `--object-id` (keeps `-n`/`--name`
  aliases); help now states it is the Entra object (client) ID GUID, not a
  friendly name or UPN.
- Make all command/group help Entra-specific; fix "defintion" typo in show.
- Update examples to use a GUID object id.
- Fix roles singular option in the aaz model (was generating ----role).
- Re-record the entra-user and negative scenario tests live; both pass playback.

The flat --type wrapper (custom.py) and its command-table registration are
rewired to the entra-user name so the flattening survives the rename.

Copilot-Session: 40a4fbc4-fd62-4bcc-9b7d-6e70f80e114d
