+++ src/documentdb/azext_documentdb/aaz/latest/documentdb/mongocluster/entra_user/_create.py
@@ -12,13 +12,13 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "documentdb mongocluster user create",
+    "documentdb mongocluster entra-user create",
 )
 class Create(AAZCommand):
-    """Create a new user or updates an existing user on a mongo cluster.
+    """Create or update a Microsoft Entra ID user on a mongo cluster.
 
-    :example: Create an Entra-backed user.
-        az documentdb mongocluster user create -n alice-entra --cluster-name MyCluster -g MyResourceGroup --type User --role db=admin role=root
+    :example: Create an Entra ID user (service principal or user) by object ID.
+        az documentdb mongocluster entra-user create --object-id 11111111-1111-1111-1111-111111111111 --cluster-name MyCluster -g MyResourceGroup --type User --role db=admin role=root
     """
 
     _aaz_info = {
@@ -58,9 +58,9 @@ class Create(AAZCommand):
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
         )
-        _args_schema.user_name = AAZStrArg(
-            options=["-n", "--name", "--user-name"],
-            help="The name of the mongo cluster user.",
+        _args_schema.object_id = AAZStrArg(
+            options=["-n", "--name", "--object-id"],
+            help="Object ID (client ID) of the Microsoft Entra principal. Provide the GUID of the service principal or user, not a friendly name or UPN.",
             required=True,
             fmt=AAZStrArgFormat(
                 pattern="^[a-zA-Z0-9\\-]*",
@@ -189,7 +189,7 @@ class Create(AAZCommand):
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "userName", self.ctx.args.user_name,
+                    "userName", self.ctx.args.object_id,
                     required=True,
                 ),
             }
