+++ src/documentdb/azext_documentdb/tests/latest/recordings/test_documentdb_mongocluster_negative.yaml
@@ -29,7 +29,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Tue, 14 Jul 2026 20:48:50 GMT
+      - Fri, 17 Jul 2026 21:23:39 GMT
       expires:
       - '-1'
       pragma:
@@ -43,7 +43,7 @@ interactions:
       x-ms-failure-cause:
       - gateway
       x-msedge-ref:
-      - 'Ref A: CF4D0CA67F6D4B4783367B7A6D7658B9 Ref B: MNZ221060608027 Ref C: 2026-07-14T20:48:50Z'
+      - 'Ref A: C4F0C8F5324B49BCBCC64C644FE1934F Ref B: BL2AA2030101049 Ref C: 2026-07-17T21:23:40Z'
     status:
       code: 404
       message: Not Found
@@ -77,7 +77,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Tue, 14 Jul 2026 20:48:50 GMT
+      - Fri, 17 Jul 2026 21:23:40 GMT
       expires:
       - '-1'
       pragma:
@@ -91,7 +91,7 @@ interactions:
       x-ms-failure-cause:
       - gateway
       x-msedge-ref:
-      - 'Ref A: E2A40F2B0BE345FBABAF74A25B76923F Ref B: MNZ221060618035 Ref C: 2026-07-14T20:48:51Z'
+      - 'Ref A: FA3C08E62F8B4F37BCB387536E9AD266 Ref B: MNZ221060619031 Ref C: 2026-07-17T21:23:40Z'
     status:
       code: 404
       message: Not Found
@@ -103,7 +103,7 @@ interactions:
       Accept-Encoding:
       - gzip, deflate
       CommandName:
-      - documentdb mongocluster user show
+      - documentdb mongocluster entra-user show
       Connection:
       - keep-alive
       ParameterSetName:
@@ -125,7 +125,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Tue, 14 Jul 2026 20:48:52 GMT
+      - Fri, 17 Jul 2026 21:23:40 GMT
       expires:
       - '-1'
       pragma:
@@ -139,7 +139,7 @@ interactions:
       x-ms-failure-cause:
       - gateway
       x-msedge-ref:
-      - 'Ref A: E6FB943759A64AEEB3C740F565529403 Ref B: BL2AA2030101039 Ref C: 2026-07-14T20:48:52Z'
+      - 'Ref A: 41AA53F41F8148F5B4D3BD50BB4D12A2 Ref B: BL2AA2010204035 Ref C: 2026-07-17T21:23:41Z'
     status:
       code: 404
       message: Not Found
@@ -173,7 +173,7 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Tue, 14 Jul 2026 20:48:52 GMT
+      - Fri, 17 Jul 2026 21:23:41 GMT
       expires:
       - '-1'
       pragma:
@@ -187,7 +187,7 @@ interactions:
       x-ms-failure-cause:
       - gateway
       x-msedge-ref:
-      - 'Ref A: E6447DB5B401400DA54ADFD107EC8F60 Ref B: BL2AA2010204027 Ref C: 2026-07-14T20:48:53Z'
+      - 'Ref A: 6CFDAF9685654844B906FB7F727A2721 Ref B: BL2AA2010205033 Ref C: 2026-07-17T21:23:42Z'
     status:
       code: 404
       message: Not Found
